dashboard-api/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2
httpx
python-dotenv
pydantic
